move card layout to layout folder
