+++ src/components/layout/ExpandedCardLayout.tsx
